Move QWebView creation to its own method and make it transparent
The slideshow played by the web view use a fixed color which approximately
matches the background texture used in the installer.

Now that the web view does not fill the entire window width anymore, the
difference between this fixed color and the background texture is becoming more
visible on the left and right edges. Making the web view transparent (and fixing
the slideshow, in a separate MR) avoids this issue.