[Autopilot tests] Add workaround for write changes to disk dialog
Unfortunately the dialog buttons have not identifyable properties to be
able to select and click the continue button. So this patch resorts to
a lengthy sleep and keyboard navigation to continue the install.

I will file a seperate bug about improving autopilot accessibility of
this dialog as it would mean a change happening in ubiquity itself which
I don't want to include in this commit