Enable a minimal install functionality
The option only appears in the GTK and Qt environments if a file
listing packages to remove is present in:
/usr/share/ubiquity/manifests/minimal_install_removal.
  
Note that this is a short-term solution. When moving to subiquity, we'll
have livefs mounts on top of each other to provide this functionality.
This is why we went to that intermediate feature for 18.04.