New minimal installation for ubuntu desktop
Add a new ubiquity-ubuntu-desktop-minimal package to offer minimal
installation option in ubuntu desktop.

Note that we need, as a workaround, listing principal rdepends and l10n
packages related to the ones we are removing to not have to relax their
dependencies pulling them back on finale installation.
This is until we move to subiquity and multiple layered rootfs.

As of today, normal installation is 4.1G and minimal is 3.5G.