gtk, kde: Start installing after ubi-timezone
When installing in a non-US English (en) locale, Ubiquity currently
forgets that you've done this for package selection purposes and
proceeds as if you'd selected en_US.

This is because, although the selection is recorded correctly by the UI,
it is not committed to debconf in time for the actual installation code
to see it. Install is started after the partman_commit stage, but we do
not know which langpacks to keep until after localechooser runs from
ubi-timezone.

Move the start of installation until after the timezone stage, so that
install.py knows which locale the user selected when it starts. This
also requires the same "parallel db" treatment as for partman-commit, so
we make a new one of those too.

https://bugs.launchpad.net/ubuntu/+source/ubiquity/+bug/1797579