fix crash in kcrash init caused by incorrect argv
kcrash will edit the qstringlist of qapp::arguments() and assume it to
have at least one entry. this would cause a crash with how we
initialized qapp.

specifically initializing with an empty list is not a valid thing to do.
the qcoreapplication documentation explicitly requires that the arg data
must stay valid the entire lifetime AND that there must be at least one
argv entry. this is in line with general posix expectations where
argv[0] is always the called executable.

seeing as we need a valid argument list passed in, we may as well pass the
correct sys.argv into the qapp.