gtk,kde: Don't start installing multiple times
If we go back and forth between timezone and usersetup multiple times,
we can end up restarting debconf under an in-progress installation
process or starting the installation multiple times. This is obviously
bad, and can end up with the installer crashing.

Only restart debconf and start the installation the *first* time the
user clicks next from the timezone step.

https://bugs.launchpad.net/ubuntu/+source/ubiquity/+bug/1799097