Support usr-merge in ubiquity.
With introduction of usr-merge in disco, we started to have top-level
symlinks to dirs i.e. /sbin -> /usr/sbin which broke with blacklisting
logic. Because such a symlink may not be blacklisted, even though
blacklisted packages ship /sbin directory. A simple check if that is a
dir is no longer sufficient. Instead extended the blacklisting logic
to prevent blacklisting symlinks to dirs.

LP: #1801408