Add experiemental support for ZFS to ubiquity
A new option to install the system with ZFS is now available as an
experimental feature on the guided partitioning page. This option is
active only if the user selects to use the entire disk. It is clearly
labelled as experimental on the UI. The script zsys-setup is then used
to create the pool and dataset compatible with zsys and the first user
migrated to its own dataset.

Co-authored-by: Didier Roche <didrocks@ubuntu.com>