nm.py: Stop forking udevadm and call GUdev to retrieve network adapter info
systemd 242-6ubuntu1 dropped a `/sbin/ubiquity` compatibility symlink
that we were referencing by absolute path. This broke ubiquity when run
on a system with a wireless adapter.

It'd be better instead to use a library to interface with udev, rather
than executing a subprocess and parsing its output. Let's modify the
code to do that instead.

LP: #1845730